+++ src/Config/config.php
@@ -3,7 +3,7 @@
 return (object) [
 	'service'=>[
         'document'=>'https://webservices.brightree.net/v0100-2108/DocumentationService/DocumentManagementService.svc',
-        'patient' => 'https://webservices.brightree.net/v0100-2101/OrderEntryService/patientservice.svc',
+        'patient' => 'https://webservices.brightree.net/v0100-2410/OrderEntryService/patientservice.svc',
         'documentation' => 'https://webservices.brightree.net/v0100-2004/DocumentationService/DocumentationService.svc',
         'custom' => 'https://webservices.brightree.net/v0100-1610/CustomFieldService/CustomFieldService.svc',
         'insurance' => 'https://webservices.brightree.net/v0100-2310/OrderEntryService/InsuranceService.svc',
@@ -11,7 +11,7 @@ return (object) [
         'doctor' => 'https://webservices.brightree.net/v0100-2406/DoctorService/DoctorService.svc',
         'inventory' => 'https://webservices.brightree.net/v0100-2309/InventoryService/InventoryService.svc',
         'pickup' => 'https://webservices.brightree.net/v0100-2007/OrderEntryService/PickupExchangeService.svc',
-        'salesorder' => 'https://webservices.brightree.net/v0100-2408/OrderEntryService/SalesOrderService.svc',
+        'salesorder' => 'https://webservices.brightree.net/v0100-2409/OrderEntryService/SalesOrderService.svc',
         'invoice' => 'https://webservices.brightree.net/v0100-2409/InvoiceService/InvoiceService.svc',
         'security' => 'https://webservices.brightree.net/v0100-2407/SecurityService/UserSecurityService.svc',
         'pricing' => 'https://webservices.brightree.net/v0100-1908/InventoryService/PricingService.svc'